requirements.txt
torch
torchvision
matplotlib
pillow
pillow-avif-plugin
pillow-heif